% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dddata-package.R
\docType{package}
\name{dddata-package}
\alias{dddata}
\alias{dddata-package}
\title{dddata: Data from the Dungeons and Dragons System Reference Document}
\description{
The System Reference Document (SRD) contains D&D rules content you can use and reference to publish content under Creative Commons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/dddata}
  \item \url{https://jonthegeek.github.io/dddata/}
  \item Report bugs at \url{https://github.com/jonthegeek/dddata/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
